chore(security-txt): update generated types and methods

+++ src/resources/security-txt.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../resource';
+import { isRequestOptions } from '../core';
 import * as Core from '../core';
 
 export class SecurityTXT extends APIResource {
@@ -16,10 +17,18 @@ export class SecurityTXT extends APIResource {
    * ```
    */
   update(
-    params: SecurityTXTUpdateParams,
+    params?: SecurityTXTUpdateParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<SecurityTXTUpdateResponse>;
+  update(options?: Core.RequestOptions): Core.APIPromise<SecurityTXTUpdateResponse>;
+  update(
+    params: SecurityTXTUpdateParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<SecurityTXTUpdateResponse> {
-    const { zone_id, ...body } = params;
+    if (isRequestOptions(params)) {
+      return this.update({}, params);
+    }
+    const { zone_id = this._client.zoneId, ...body } = params;
     return this._client.put(`/zones/${zone_id}/security-center/securitytxt`, { body, ...options });
   }
 
@@ -35,10 +44,18 @@ export class SecurityTXT extends APIResource {
    * ```
    */
   delete(
-    params: SecurityTXTDeleteParams,
+    params?: SecurityTXTDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<SecurityTXTDeleteResponse>;
+  delete(options?: Core.RequestOptions): Core.APIPromise<SecurityTXTDeleteResponse>;
+  delete(
+    params: SecurityTXTDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<SecurityTXTDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete({}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return this._client.delete(`/zones/${zone_id}/security-center/securitytxt`, options);
   }
 
@@ -53,8 +70,16 @@ export class SecurityTXT extends APIResource {
    * });
    * ```
    */
-  get(params: SecurityTXTGetParams, options?: Core.RequestOptions): Core.APIPromise<SecurityTXTGetResponse> {
-    const { zone_id } = params;
+  get(params?: SecurityTXTGetParams, options?: Core.RequestOptions): Core.APIPromise<SecurityTXTGetResponse>;
+  get(options?: Core.RequestOptions): Core.APIPromise<SecurityTXTGetResponse>;
+  get(
+    params: SecurityTXTGetParams | Core.RequestOptions = {},
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<SecurityTXTGetResponse> {
+    if (isRequestOptions(params)) {
+      return this.get({}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.get(`/zones/${zone_id}/security-center/securitytxt`, options) as Core.APIPromise<{
         result: SecurityTXTGetResponse;
@@ -177,7 +202,7 @@ export interface SecurityTXTUpdateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param
@@ -229,14 +254,14 @@ export interface SecurityTXTDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface SecurityTXTGetParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export declare namespace SecurityTXT {
